requirements.txt
# Core
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0

# Trading & OANDA
oandapyV20
python-dotenv>=0.21.0
ccxt>=2.6.85

# Machine Learning
torch>=1.13.0
stable-baselines3>=1.7.0
tensorboard>=2.10.0

# Additional
redis>=4.3.4
fastapi>=0.85.0  # For monitoring dashboard
uvicorn>=0.19.0